requirements.txt
# Core dependencies
langchain==0.3.7
langchain-openai==0.2.6
redis==5.0.1

# API and monitoring
fastapi==0.109.0
uvicorn==0.27.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
